Pull editor opening and line selecting code out of PydevConsoleCommunication
Editor opening now in PyOpenEditor
Line selecting now in EditorUtils
